Add files via upload
This project helps you to upload a photo or a video, and check whether it is a genuine digital data or a Deep-Fake. This 2D Convolutional Neural Network improves the accuracy and makes the model more robust in its performance. This extracts high level features in from the input and breaks down the features of a face in the given photo or a video and determines if it is a Real or a Deep-Fake.